+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ### Breaking changes
 * Soft deprecate function `label_aevar()`.
-
+* Replaced `chunks` with simpler `Quosure` class.
+* Replaced `datasets` argument containing `FilteredData` with the new arguments `data` (list of reactive datasets) and `filter_panel_api` (`FilterPanelAPI`).
+* Updated `arm_var` to point to the factor column in `ANL`. It can't be a character column anymore. 
 
 ### Enhancements
 * Added `teal.reporter` to all modules.